Configure linting and formatting
yarn fmt -- auto-formats

move App.tsx to `src` so it will be
easier for us to configure the location of the ts project.

And move App test to App.test.tsx - for tests that
exercise a single module, it is more convenient to have
the test next to the code.