enable editing plants (#29)
close #15

## Description

Tap a plant on the main screen to get to the "Plant" screen.

This screen has a button for editing the plant. We can also add buttons here in future for:
- watering (calling)
- removing

Tap "Edit Plant" to update the species and/or watering frequency.

## Notes

Note: I wasn't sure what to call the new test file–ideas welcome. `plant-screen-test.tsx` may not fit with our current conventions.

This PR also includes refactors:
- extract part of HomeScreen into a PlantView component. This is now shared between the Plant screen and the Home screen. In future, we ma want to display these slightly differently, but was convenient for this PR.
- change the `addPlant` test util to `setPlantOptions`. This is shared code for setting watering frequency and species, it is used by both `add-plants-test.tsx` and `plants-screen-test.tsx`.
- the test 'get items from async storage' is now deterministic - we now mock the date.